+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ The format is based on `Keep a Changelog <http://keepachangelog.com/>`_.
 22.10.18 - 22.10.18
 =====================
 * Added csv for Monitoring: agents, events, alarm, db management and Notifications
+* Added Event Actions
 
 =====================
 22.10.11 - 22.10.11